+++ src/main/java/io/gravitee/gateway/reactive/api/context/http/HttpPlainExecutionContext.java
@@ -17,8 +17,10 @@ package io.gravitee.gateway.reactive.api.context.http;
 
 import io.gravitee.gateway.api.buffer.Buffer;
 import io.gravitee.gateway.reactive.api.ExecutionFailure;
+import io.gravitee.gateway.reactive.api.policy.base.BasePolicy;
 import io.reactivex.rxjava3.core.Completable;
 import io.reactivex.rxjava3.core.Maybe;
+import java.util.function.Function;
 
 /**
  * Http execution context allowing access to {@link HttpPlainRequest} and {@link HttpPlainResponse} that manipulates raw http bodies (content or chunks).
@@ -65,4 +67,15 @@ public interface HttpPlainExecutionContext extends HttpBaseExecutionContext {
      * Same as {@link #interruptBody()} but with an {@link ExecutionFailure} object that indicates that the execution has failed. The {@link ExecutionFailure} can be processed in order to build a proper response (ex: based on templating, with appropriate accept-encoding, ...).
      */
     Maybe<Buffer> interruptBodyWith(final ExecutionFailure failure);
+
+    /**
+     * Registers an action defined during the request phase to be executed later during the response phase.
+     * When multiple actions are registered, they are executed in reverse order of registration (LIFO).
+     *
+     * @param source the policy that initiated the action.
+     * @param onResponseCallback the action to be executed during the response phase, receiving the HTTP execution context and returning a Completable.
+     */
+    default void addActionOnResponse(BasePolicy source, Function<HttpExecutionContext, Completable> onResponseCallback) {
+        // Nothing is registered by default
+    }
 }