+++ CHANGELOG.md
@@ -1,3 +1,95 @@
+# [4.1.0-alpha.13](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.12...4.1.0-alpha.13) (2025-12-11)
+
+
+### Features
+
+* add connection id in KafkaConnectionContext ([e1d6663](https://github.com/gravitee-io/gravitee-gateway-api/commit/e1d66639edefe4f3b78a4cf0872002f191c69538))
+
+# [4.1.0-alpha.12](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.11...4.1.0-alpha.12) (2025-12-11)
+
+
+### Features
+
+* add A2A Proxy specialized connector interfaces ([188b637](https://github.com/gravitee-io/gravitee-gateway-api/commit/188b637767749436e0b60945badb638a11dbe3ad))
+
+# [4.1.0-alpha.11](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.10...4.1.0-alpha.11) (2025-12-09)
+
+
+### Features
+
+* add the BaseSecurityPolicy.requireSubscription(BaseExecutionContext context) method ([5309c2a](https://github.com/gravitee-io/gravitee-gateway-api/commit/5309c2abd88fe06cec22b32fe907a4cfef9780bb))
+
+# [4.1.0-alpha.10](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.9...4.1.0-alpha.10) (2025-12-05)
+
+
+### Bug Fixes
+
+* identify the source of the action to execute on response ([9d675f5](https://github.com/gravitee-io/gravitee-gateway-api/commit/9d675f513506137dab8f154480597e35946eb20c))
+
+# [4.1.0-alpha.9](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.8...4.1.0-alpha.9) (2025-12-01)
+
+
+### Bug Fixes
+
+* add default implementation ([5d1b39c](https://github.com/gravitee-io/gravitee-gateway-api/commit/5d1b39cff1641580cac9dc65e2ed3f6611229737))
+
+# [4.1.0-alpha.8](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.7...4.1.0-alpha.8) (2025-12-01)
+
+
+### Features
+
+* add the wwwAuthenticate() and onWellKnown() default methods to HttpSecurityPolicy ([7c57b48](https://github.com/gravitee-io/gravitee-gateway-api/commit/7c57b48d82c051b8c8d8b7e88a58818f5deeae8e))
+
+# [4.1.0-alpha.7](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.6...4.1.0-alpha.7) (2025-11-25)
+
+
+### Features
+
+* add attribute for tracing with verbose ([b949df6](https://github.com/gravitee-io/gravitee-gateway-api/commit/b949df68545c4f375a5ee403f0bf4ebebc308b05))
+
+# [4.1.0-alpha.6](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.5...4.1.0-alpha.6) (2025-11-17)
+
+
+### Features
+
+* add additional internal attributes ([dac33f5](https://github.com/gravitee-io/gravitee-gateway-api/commit/dac33f50585941f0ca97206fc165103cc4c15827))
+
+# [4.1.0-alpha.5](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.4...4.1.0-alpha.5) (2025-11-04)
+
+
+### Features
+
+* create a new internal attribute to override Content-Type filtering in logs ([4af927b](https://github.com/gravitee-io/gravitee-gateway-api/commit/4af927b5c93fc4be6ed2a06462895a8f42a2c40c))
+
+# [4.1.0-alpha.4](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.3...4.1.0-alpha.4) (2025-10-27)
+
+
+### Features
+
+* add action that will be executed at http request response phase ([68e499c](https://github.com/gravitee-io/gravitee-gateway-api/commit/68e499caaaf608684a3dac4d393f257728e7f304))
+
+# [4.1.0-alpha.3](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.2...4.1.0-alpha.3) (2025-10-17)
+
+
+### Features
+
+* create new LLM proxy interfaces ([69f5b88](https://github.com/gravitee-io/gravitee-gateway-api/commit/69f5b88c22211dbbbdca3c9635cf1d298043ee0d))
+
+# [4.1.0-alpha.2](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.1.0-alpha.1...4.1.0-alpha.2) (2025-10-10)
+
+
+### Bug Fixes
+
+* ignore cause in ExecutionFailure.equals ([935f0d9](https://github.com/gravitee-io/gravitee-gateway-api/commit/935f0d9cde1b1f2e267626c5dafa2113af058a12))
+
+# [4.1.0-alpha.1](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.0.1...4.1.0-alpha.1) (2025-10-06)
+
+
+### Features
+
+* add llm-proxy api type ([454a365](https://github.com/gravitee-io/gravitee-gateway-api/commit/454a3650b81f541d96ce12e9e36bb3d67e854e81))
+* add mcp-proxy api type ([6ececc5](https://github.com/gravitee-io/gravitee-gateway-api/commit/6ececc5b949f8a280572cfe34b37231389c3c9ee))
+
 ## [4.0.2](https://github.com/gravitee-io/gravitee-gateway-api/compare/4.0.1...4.0.2) (2025-10-10)